requirements.txt
astroid==2.5.6
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
parse==1.19.0
pycodestyle==2.7.0
pygitea==0.0
pylint==2.8.3
python-dateutil==2.8.1
python-dotenv==0.18.0
requests==2.25.1
six==1.16.0
smmap==4.0.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
wrapt==1.12.1
